Adição editar e checar Cargo

+++ BancoAdministrador.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 
 import dao.AdministradorDAO;
 import model.entity.Administrador;
+import model.entity.AdministradorAtributos;
+import model.entity.AdministradorCargo;
 
 public class BancoAdministrador {
 	
@@ -18,23 +20,70 @@ public class BancoAdministrador {
 		administradores = new ArrayList<Administrador>();
 	}
 	
-	public void adicionar(Administrador administrador) {
-		// Id gerenciado pelo programa
-		administrador.setId(idAtual++);
-		administradores.add(administrador);
-		AdministradorDAO.salvar(administrador);
+	public boolean adicionar(Administrador administrador, Object user) {
 		
-		quantidade++;
-	}
-	
-	public boolean remover(Administrador administrador) {
-		if(administradores.contains(administrador)) {
-			administradores.remove(administrador);
-			AdministradorDAO.remover(administrador.getId());
+		if(checarPermissao(user)) {
+			// Id gerenciado pelo programa
+			administrador.setId(idAtual++);
+			administradores.add(administrador);
+			AdministradorDAO.salvar(administrador);
 			
-			quantidade --;
+			quantidade++;
 			
 			return true;
+		}
+		return false;		
+	}
+	
+	public boolean remover(Administrador administrador, Object user) {
+		if(checarPermissao(user)) {
+			if(administradores.contains(administrador)) {
+				administradores.remove(administrador);
+				AdministradorDAO.remover(administrador.getId());
+				
+				quantidade --;
+				
+				return true;
+			}
+		}
+		
+		return false;
+	}
+	
+	public boolean editar(AdministradorAtributos atributo, Administrador administrador, Object valor, Object user) {
+		if(checarPermissao(user)) {
+			if(administradores.contains(administrador)) {
+				
+				switch(atributo) {
+				
+				case AdministradorAtributos.nome:
+					administrador.setNome((String) valor);
+					break;
+				case AdministradorAtributos.cpf:
+					administrador.setCpf((String) valor);
+					break;
+				
+				case AdministradorAtributos.senha:
+					administrador.setSenha((String) valor);
+					break;
+				
+				case AdministradorAtributos.cargo:
+					administrador.setCargo((AdministradorCargo) valor);
+					break;
+				
+				default:
+					System.err.println("Passe um Atributo válido");
+				
+				//Troca os administradores
+				administradores.set(administradores.indexOf(administrador), administrador);
+				AdministradorDAO.editar(atributo, (String) valor, administrador.getId());
+				
+				return true;
+			}
+			
+			return false;
+			
+			}
 		}
 		
 		return false;
@@ -59,6 +108,20 @@ public class BancoAdministrador {
 		}
 		
 		return resultado.toString();	
+	}
+	
+	public boolean checarPermissao(Object user) {
+		if(user instanceof Administrador) {
+			Administrador adm = (Administrador) user;
+			
+			if(adm.getCargo() == AdministradorCargo.GERENTE) {
+				return true;
+			}
+			
+			return false;	
+		}
+		
+		return false;
 	}
 	
 	// Só pra facilitar o desenvolvimento caso esqueça que o toString tbm lista kk
